src/prime: improve method getting divisors of N.
Improve a little the method filling a vector with all divisors of N
(included N).

+++ src/math/prime.c
@@ -145,12 +145,12 @@ get_all_primes_factors_of_n(unsigned long n, const gv_t(uint64) *primes,
 
 static void
 get_all_divisors_of_n_from_prime_factors_rec(const gv_t(primes_factors) *in,
-                                             long basis, long first_pos,
-                                             gv_t(uint64) *out)
+                                             unsigned long basis,
+                                             long first_pos, gv_t(uint64) *out)
 {
-    for (int pos = first_pos; pos < in->len; pos++) {
+    for (long pos = first_pos; pos < in->len; pos++) {
         prime_factor_t *current_prime_factor = &(in->tab[pos]);
-        long current_div = current_prime_factor->prime * basis;
+        unsigned long current_div = current_prime_factor->prime * basis;
 
         for (unsigned int i = 0; i < current_prime_factor->iteration; i++) {
             gv_add(out, current_div);
@@ -164,11 +164,12 @@ get_all_divisors_of_n_from_prime_factors_rec(const gv_t(primes_factors) *in,
 void get_all_divisors_of_n_from_prime_factors(const gv_t(primes_factors) *in,
                                               gv_t(uint64) *out)
 {
+    /* 1 divides all the numbers */
     gv_add(out, 1);
 
     gv_for_each_pos(pos, in) {
         prime_factor_t *current_prime_factor = &(in->tab[pos]);
-        long current_div = current_prime_factor->prime;
+        unsigned long current_div = current_prime_factor->prime;
 
         for (unsigned int i = 0; i < current_prime_factor->iteration; i++) {
             gv_add(out, current_div);
